+++ example_backend/examples/authoritative_rts.rs
@@ -202,6 +202,7 @@ fn apply_unit_spawn(
 
     commands.spawn((
         Unit,
+        Replicated,
         team,
         Transform::from_translation(spawn.position.extend(0.0)),
     ));
@@ -685,7 +686,7 @@ struct MoveUnits {
 
 #[derive(Component, Serialize, Deserialize)]
 #[component(immutable)]
-#[require(Replicated, Command, Mesh2d, MeshMaterial2d<ColorMaterial>)]
+#[require(Command, Mesh2d, MeshMaterial2d<ColorMaterial>)]
 struct Unit;
 
 #[derive(Component)]